Commit minor changes, moving main local repo to spin directory with more storage space
